feat(plutus): strip primary app to settlement accounting
