+++ apps/plutus/app/bills/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
 export default function BillsPage() {
-  redirect('/transactions?tab=bill');
+  redirect('/cogs-inputs');
 }
