+++ apps/plutus/app/cogs-inputs/page.tsx
@@ -1,0 +1,3 @@
+import { CogsInputsPage } from '@/components/cogs-inputs/cogs-inputs-page';
+
+export default CogsInputsPage;
